refactoring of services and controllers, added postman collection

+++ tourFinder/src/main/java/fmi/java/web/tourFinder/service/UserService.java
@@ -17,26 +17,25 @@ public class UserService {
 
     public Validation<List<String>, User> create(User user) {
         var validation = UserValidator.validate(user);
-        if (validation.isValid()) {
-
-            var existingUsername = checkForExistingUsername(user);
-
-            var existingEmail = checkForExistingEmail(user);
-
-            if (!existingUsername.isValid() || !existingEmail.isValid()) {
-                var errors = new ArrayList<String>();
-
-                if (existingUsername.getErrors() != null) {
-                    errors.add(existingUsername.getErrors());
-                }
-                if (existingEmail.getErrors() != null) {
-                    errors.add(existingEmail.getErrors());
-                }
-                return Validation.invalid(errors);
-            }
-            userRepository.save(user);
+        if (!validation.isValid()) {
+            return validation;
         }
-        return validation;
+        var existingUsername = checkForExistingUsername(user);
+
+        var existingEmail = checkForExistingEmail(user);
+
+        if (!existingUsername.isValid() || !existingEmail.isValid()) {
+            var errors = new ArrayList<String>();
+
+            if (existingUsername.getErrors() != null) {
+                errors.add(existingUsername.getErrors());
+            }
+            if (existingEmail.getErrors() != null) {
+                errors.add(existingEmail.getErrors());
+            }
+            return Validation.invalid(errors);
+        }
+        return Validation.valid(userRepository.save(user));
     }
 
     private Validation<String, User> checkForExistingUsername(User user) {
@@ -58,13 +57,45 @@ public class UserService {
     }
 
     public User getById(String id) {
-        return userRepository.findById(id).orElse(null);
+        return userRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("User with id " + id + " does not exist"));
     }
 
-    public void update(User user, String id) {
+    public Validation<List<String>, User> update(User user, String id) {
+        var validation = UserValidator.validate(user);
+        if (!validation.isValid()) {
+            return validation;
+        }
+
+        var existingUsername = checkForExistingUsername(user);
+
+        var existingEmail = checkForExistingEmail(user);
+
+        if (!existingUsername.isValid() || !existingEmail.isValid()) {
+            var errors = new ArrayList<String>();
+
+            if (existingUsername.getErrors() != null) {
+                errors.add(existingUsername.getErrors());
+            }
+            if (existingEmail.getErrors() != null) {
+                errors.add(existingEmail.getErrors());
+            }
+            return Validation.invalid(errors);
+        }
+
+        if (userRepository.findById(id).isPresent()) {
+            return Validation.valid(userRepository.save(user));
+        } else {
+            var errors = new ArrayList<String>();
+            errors.add("User with id " + id + " does not exist");
+            return Validation.invalid(errors);
+        }
     }
 
     public void delete(String id) {
-        userRepository.deleteById(id);
+        if (userRepository.findById(id).isPresent()) {
+            userRepository.deleteById(id);
+        } else {
+            throw new IllegalArgumentException("User with id " + id + " does not exist");
+        }
     }
 }
